Rename file stage-and-actors.mjs into motion-stage.mjs.
