+++ lib/js/components/axes-math.mjs
@@ -44,7 +44,7 @@ import {
 
 import {
     SelectAndDragByOptions
-} from './layouts/stage-and-actors.mjs';
+} from './layouts/motion-stage.mjs';
 
 import {
     binarySearch
