+++ lib/js/components/dev-layouts/videoproof-array.mjs
@@ -48,7 +48,7 @@ import {
 import {
     initAnimanion
   , UITimeControl
-} from '../layouts/stage-and-actors.mjs';
+} from '../layouts/motion-stage.mjs';
 
 
 const VideoproofArrayActorModel = activatableActorTypes.get('VideoproofArrayActorModel').get('typeClass').value;
@@ -899,13 +899,13 @@ class VideoproofArrayController extends _BaseContainerComponent {
               , UICharGroupContainer
               , zones
               , { // injectable
-                    // not implemented: _getArgumentConfig http://localhost:8080/lib/js/components/layouts/stage-and-actors.mjs:2771
+                    // not implemented: _getArgumentConfig http://localhost:8080/lib/js/components/layouts/motion-stage.mjs:2771
                     updateDefaultsDependencies
-                    // not implemented: _getArgumentConfig http://localhost:8080/lib/js/components/layouts/stage-and-actors.mjs:2775
+                    // not implemented: _getArgumentConfig http://localhost:8080/lib/js/components/layouts/motion-stage.mjs:2775
                   , requireUpdateDefaults
                   , genericTypeToUIElement
-                    // get: not implemented: UICharGroupContainer http://localhost:8080/lib/js/components/layouts/stage-and-actors.mjs:2955
-                    // use: not implemented:  _activateCustom http://localhost:8080/lib/js/components/layouts/stage-and-actors.mjs:3020
+                    // get: not implemented: UICharGroupContainer http://localhost:8080/lib/js/components/layouts/motion-stage.mjs:2955
+                    // use: not implemented:  _activateCustom http://localhost:8080/lib/js/components/layouts/motion-stage.mjs:3020
                     // Uncaught (in promise) Error: not implemented: get getDefaults(prefix:string:generic/charGroup, key:string:options, defaultVal:object:null)
                   , getDefaults: function(ppsRecord, modelFieldName, defaultVal=_NOTDEF) {
                         // This is similar to KeyMomentController._getDefaults
