+++ lib/js/components/layouts/motion-stage.mjs
