+++ lib/js/components/main-ui.mjs
@@ -30,7 +30,7 @@ import {
 import * as ExampleLayout from './dev-layouts/example.mjs';
 import * as ExampleKeyMomentsLayout from './dev-layouts/example-key-moments.mjs';
 import * as PentrekLayout from './dev-layouts/pentrek.mjs';
-import * as MotionStageLayout from './layouts/stage-and-actors.mjs';
+import * as MotionStageLayout from './layouts/motion-stage.mjs';
 import * as VideoproofArrayLayout from './dev-layouts/videoproof-array.mjs';
 import * as VideoproofLayout from './layouts/videoproof.typeroof.jsx';
 import * as TypeStage from './layouts/type-stage/index.typeroof.jsx';
