+++ lib/js/main-player.mjs
@@ -31,7 +31,7 @@ import {
     IDRegistry
 } from './components/dom-id-registry.mjs';
 
-import {PlayerLayout as MotionStageLayout} from './components/layouts/stage-and-actors.mjs';
+import {PlayerLayout as MotionStageLayout} from './components/layouts/motion-stage.mjs';
 import * as VideoproofArrayLayout from './components/dev-layouts/videoproof-array.mjs';
 import * as VideoproofLayout from './components/layouts/videoproof.typeroof.jsx';
 import * as TypeStage from './components/layouts/type-stage/index.typeroof.jsx';
